Use TargetManagerInterface which allows to use a NullTargetManager in case loading a file stream fails
